+++ ambry-replication/src/test/java/com.github.ambry.replication/MockFindToken.java
@@ -14,6 +14,7 @@
 package com.github.ambry.replication;
 
 import com.github.ambry.store.FindToken;
+import com.github.ambry.store.FindTokenFactory;
 import java.io.DataInputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -59,5 +60,18 @@ public class MockFindToken implements FindToken {
 
   public long getBytesRead() {
     return this.bytesRead;
+  }
+
+  public static class MockFindTokenFactory implements FindTokenFactory {
+
+    @Override
+    public FindToken getFindToken(DataInputStream stream) throws IOException {
+      return new MockFindToken(stream);
+    }
+
+    @Override
+    public FindToken getNewFindToken() {
+      return new MockFindToken(0, 0);
+    }
   }
 }
